+++ src/index.html
@@ -21,15 +21,19 @@
     <!-- ❗️❗️❗️ Add partials like so for every page -->
     <include src="./partials/header.html"></include>
     <div class="pagination"></div>
-    <include src="./partials/modal.html"></include>
+  <!-- <div id="data-container"></div>
+    <div id="pagination-container"></div> -->
 
-    <!-- <div id="pagination-container"></div> -->
+
+    <include src="./partials/modal.html"></include>
 
     <include src="./partials/gallery.html"></include>
     <include src="./partials/footer.html"></include>
 
     <script type="module" src="./js/config.js"></script>
     <script type="module" src="./index.js"></script>
+    <script type="module" src="./js/renderEvents.js"></script>
+    <script type="module" src="./js/renderPagination.js"></script>
     <script type="module" src="./js/render.js"></script>
     <script type="module" src="./js/modal.js"></script>
     <!-- <script type="module" src="./js/pagination.js"></script> -->
