+++ src/partials/modal.html
@@ -5,4 +5,4 @@
     </button>
   </div>
 </div>
-<include src="./partials/gallery.html"></include>
+<!-- <include src="./partials/gallery.html"></include> -->
