+++ src/sass/index.scss
@@ -9,5 +9,6 @@
 @import 'components/footer';
 @import 'components/gallery';
 @import 'components/modal';
+@import 'components/pagination';
 @import 'layout/container';
 @import 'layout/section';
